+++ src/Scenario.test.tsx
@@ -1,9 +1,9 @@
 import React from 'react'
 import ReactDOM from 'react-dom'
-import App from './App'
+import Scenario from './Scenario'
 
 it('renders without crashing', () => {
   const div = document.createElement('div')
-  ReactDOM.render(<App />, div)
+  ReactDOM.render(<Scenario />, div)
   ReactDOM.unmountComponentAtNode(div)
 })
